refactor(profile/page) replace new-profile-form with updated profile-form
